+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Add Radio and Toggle stories - [#41](https://github.com/PrefectHQ/miter-design/pull/41)
 - Add Skeleton directive - [#37](https://github.com/PrefectHQ/miter-design/pull/37)
 - Add a pop up component - [#30](https://github.com/PrefectHQ/miter-design/pull/30)
 - Add Select component - [#25](https://github.com/PrefectHQ/miter-design/pull/25)